Fix few warnings after enabling analyzers everywhere
