+++ jobs/Backend/Task/ExchangeRateUpdater.Tests/.editorconfig
@@ -1,0 +1,3 @@
+﻿[*.cs]
+
+dotnet_diagnostic.CA1861.severity = none # CA1861: Avoid constant arrays as arguments
